+++ cmd/controller/api/routes.go
@@ -19,6 +19,7 @@ package api
 
 import (
 	"net/http"
+	"github.com/vmware/purser/cmd/controller/api/apiHandlers"
 )
 
 // Route structure
@@ -37,198 +38,198 @@ var routes = Routes{
 		"GetHomePage",
 		"GET",
 		"/api",
-		GetHomePage,
+		apiHandlers.GetHomePage,
 	},
 	Route{
 		"GetPodInteractions",
 		"GET",
 		"/api/interactions/pod",
-		GetPodInteractions,
+		apiHandlers.GetPodInteractions,
 	},
 	Route{
 		"GetClusterHierarchy",
 		"GET",
 		"/api/hierarchy",
-		GetClusterHierarchy,
+		apiHandlers.GetClusterHierarchy,
 	},
 	Route{
 		"GetNamespaceHierarchy",
 		"GET",
 		"/api/hierarchy/namespace",
-		GetNamespaceHierarchy,
+		apiHandlers.GetNamespaceHierarchy,
 	},
 	Route{
 		"GetDeploymentHierarchy",
 		"GET",
 		"/api/hierarchy/deployment",
-		GetDeploymentHierarchy,
+		apiHandlers.GetDeploymentHierarchy,
 	},
 	Route{
 		"GetReplicasetHierarchy",
 		"GET",
 		"/api/hierarchy/replicaset",
-		GetReplicasetHierarchy,
+		apiHandlers.GetReplicasetHierarchy,
 	},
 	Route{
 		"GetStatefulsetHierarchy",
 		"GET",
 		"/api/hierarchy/statefulset",
-		GetStatefulsetHierarchy,
+		apiHandlers.GetStatefulsetHierarchy,
 	},
 	Route{
 		"GetPodHierarchy",
 		"GET",
 		"/api/hierarchy/pod",
-		GetPodHierarchy,
+		apiHandlers.GetPodHierarchy,
 	},
 	Route{
 		"GetContainerHierarchy",
 		"GET",
 		"/api/hierarchy/container",
-		GetContainerHierarchy,
+		apiHandlers.GetContainerHierarchy,
 	},
 	Route{
 		"GetProcessHierarchy",
 		"GET",
 		"/api/hierarchy/process",
-		GetEmptyHierarchy,
+		apiHandlers.GetEmptyHierarchy,
 	},
 	Route{
 		"GetNodeHierarchy",
 		"GET",
 		"/api/hierarchy/node",
-		GetNodeHierarchy,
+		apiHandlers.GetNodeHierarchy,
 	},
 	Route{
 		"GetPVHierarchy",
 		"GET",
 		"/api/hierarchy/pv",
-		GetPVHierarchy,
+		apiHandlers.GetPVHierarchy,
 	},
 	Route{
 		"GetPVCHierarchy",
 		"GET",
 		"/api/hierarchy/pvc",
-		GetEmptyHierarchy,
+		apiHandlers.GetEmptyHierarchy,
 	},
 	Route{
 		"GetDaemonsetHierarchy",
 		"GET",
 		"/api/hierarchy/daemonset",
-		GetDaemonsetHierarchy,
+		apiHandlers.GetDaemonsetHierarchy,
 	},
 	Route{
 		"GetJobHierarchy",
 		"GET",
 		"/api/hierarchy/job",
-		GetJobHierarchy,
+		apiHandlers.GetJobHierarchy,
 	},
 	Route{
 		"GetClusterMetrics",
 		"GET",
 		"/api/metrics",
-		GetClusterMetrics,
+		apiHandlers.GetClusterMetrics,
 	},
 	Route{
 		"GetNamespaceMetrics",
 		"GET",
 		"/api/metrics/namespace",
-		GetNamespaceMetrics,
+		apiHandlers.GetNamespaceMetrics,
 	},
 	Route{
 		"GetDeploymentMetrics",
 		"GET",
 		"/api/metrics/deployment",
-		GetDeploymentMetrics,
+		apiHandlers.GetDeploymentMetrics,
 	},
 	Route{
 		"GetDaemonsetMetrics",
 		"GET",
 		"/api/metrics/daemonset",
-		GetDaemonsetMetrics,
+		apiHandlers.GetDaemonsetMetrics,
 	},
 	Route{
 		"GetJobMetrics",
 		"GET",
 		"/api/metrics/job",
-		GetJobMetrics,
+		apiHandlers.GetJobMetrics,
 	},
 	Route{
 		"GetStatefulsetMetrics",
 		"GET",
 		"/api/metrics/statefulset",
-		GetStatefulsetMetrics,
+		apiHandlers.GetStatefulsetMetrics,
 	},
 	Route{
 		"GetReplicasetMetrics",
 		"GET",
 		"/api/metrics/replicaset",
-		GetReplicasetMetrics,
+		apiHandlers.GetReplicasetMetrics,
 	},
 	Route{
 		"GetNodeMetrics",
 		"GET",
 		"/api/metrics/node",
-		GetNodeMetrics,
+		apiHandlers.GetNodeMetrics,
 	},
 	Route{
 		"GetPodMetrics",
 		"GET",
 		"/api/metrics/pod",
-		GetPodMetrics,
+		apiHandlers.GetPodMetrics,
 	},
 	Route{
 		"GetContainerMetrics",
 		"GET",
 		"/api/metrics/container",
-		GetContainerMetrics,
+		apiHandlers.GetContainerMetrics,
 	},
 	Route{
 		"GetPVMetrics",
 		"GET",
 		"/api/metrics/pv",
-		GetPVMetrics,
+		apiHandlers.GetPVMetrics,
 	},
 	Route{
 		"GetPVCMetrics",
 		"GET",
 		"/api/metrics/pvc",
-		GetPVCMetrics,
+		apiHandlers.GetPVCMetrics,
 	},
 	Route{
 		"GetPodDiscoveryNodes",
 		"GET",
 		"/api/nodes",
-		GetPodDiscoveryNodes,
+		apiHandlers.GetPodDiscoveryNodes,
 	},
 	Route{
 		"GetPodDiscoveryEdges",
 		"GET",
 		"/api/edges",
-		GetPodDiscoveryEdges,
+		apiHandlers.GetPodDiscoveryEdges,
 	},
 	Route{
 		"GetGroupsData",
 		"GET",
 		"/api/groups",
-		GetGroupsData,
+		apiHandlers.GetGroupsData,
 	},
 	Route{
 		"Login",
 		"POST",
 		"/auth/login",
-		LoginUser,
+		apiHandlers.LoginUser,
 	},
 	Route{
 		"Logout",
 		"POST",
 		"/auth/logout",
-		LogoutUser,
+		apiHandlers.LogoutUser,
 	},
 	Route{
 		"ChangePassword",
 		"POST",
 		"/auth/changePassword",
-		ChangePassword,
+		apiHandlers.ChangePassword,
 	},
 }
